feat: use mapped method for error 409 Conflict
The idea is that prior performing the expected function of the API, we check if there is no any blocking point to perform the SQL query.

The typical situation is when a record to delete is linked to a foreign key: deleting that record will generate an error 500, which is bad because it just returns a faulty situation to the end-user. It is much better to get the list of these foreign keys and display them to the end-users so that he could understand why he cannot delete that record for the time being.

I added a field for the type of conflict, because it may have different kind of conflict. Per example a action could not be performed because an item does not have the required status.

All of this is returned with the standard http code 409.